LayoutTests/imported/w3c/web-platform-tests/editing/plaintext-only/insertHTML_white-space=pre-line-expected.txt

AC

PASS execCommand("insertHTML", false, "plaintext") when ""
PASS execCommand("insertHTML", false, "1st line
2nd line") when ""
PASS execCommand("insertHTML", false, "<pre>1st line
2nd line</pre>") when ""
FAIL execCommand("insertHTML", false, "plaintext") when "<b>{}</b>" assert_equals: expected "<b>plaintext</b>" but got "plaintext<b></b>"
FAIL execCommand("insertHTML", false, "plaintext") when "<b>{}<br></b>" assert_in_array: value "plaintext<b><br></b>" not in array ["<b>plaintext</b>", "<b>plaintext<br></b>"]
FAIL execCommand("insertHTML", false, "plaintext") when "<b>A[]B</b>" assert_equals: expected "<b>AplaintextB</b>" but got "<b>A</b>plaintext<b>B</b>"
FAIL execCommand("insertHTML", false, "plaintext") when "<span>A[]B</span>" assert_equals: expected "<span>AplaintextB</span>" but got "<span>A</span>plaintext<span>B</span>"
PASS execCommand("insertHTML", false, "<div>abc</div>") when ""
PASS execCommand("insertHTML", false, "<div>abc<br>def</div>") when ""
PASS execCommand("insertHTML", false, "<div>abc</div><div>def</div>") when ""
PASS execCommand("insertHTML", false, "<div>abc<br>def</div><div>ghi<br>jkl</div>") when ""
FAIL execCommand("insertHTML", false, "<noscript>no script</noscript>") when "" assert_equals: expected "" but got "no script"
PASS execCommand("insertHTML", false, "<noframes>no frames</noframes>") when ""
PASS execCommand("insertHTML", false, "<script>script</script>") when ""
PASS execCommand("insertHTML", false, "<style>style</style>") when ""
PASS execCommand("insertHTML", false, "<html><head><title>title</title></head><body>body</body></html>") when ""
PASS execCommand("insertHTML", false, "plain  text") when ""
PASS execCommand("insertHTML", false, "<span>plain  text</span>") when ""
FAIL execCommand("insertHTML", false, "<pre>plain  text</pre>") when "" assert_in_array: value "plain  text" not in array ["plain &nbsp;text", "plain&nbsp; text", "plain&nbsp;&nbsp;text", "plain  text", "plain  text", "plain  text"]
PASS execCommand("insertHTML", false, "") when "A[B]C"
